🔨 Refactoring: posts 페이지와  posting 페이지 분리
