+++ src/App.tsx
@@ -5,7 +5,7 @@ import Layout from './pages/layout';
 import Profile from './pages/profile';
 import SignUp from './pages/signup';
 import LogIn from './pages/login';
-import Post from './pages/post';
+import Posting from './pages/posting';
 import Meditation from './pages/meditation';
 import { queryClient, QueryClientProvider } from './apis/queryClient';
 
@@ -25,8 +25,8 @@ function App() {
                 element={<Profile />}
               />
               <Route
-                path='/posts'
-                element={<Post />}
+                path='/posting'
+                element={<Posting />}
               />
             </Route>
             <Route
