--- src/pages/post/Post.tsx
@@ -1,9 +1,0 @@
-import { Posting } from './components/Posting'
-
-const Post = () => {
-  return (
-    <Posting />
-  )
-}
-
-export default Post;
