--- src/pages/post/index.ts
@@ -1,3 +1,0 @@
-import Post from './Post';
-
-export default Post;
