+++ src/pages/posting/components/NewPost.style.ts
