+++ src/pages/posting/constants/index.ts
